Flow to river.

+++ backend/exchange/src/main/java/ar/com/tandilweb/exchange/gameProtocol/texasHoldem/inGame/DecisionInform.java
@@ -5,7 +5,7 @@ import ar.com.tandilweb.exchange.gameProtocol.texasHoldem.InGameSchema;
 public class DecisionInform extends InGameSchema {
 	
 	public String action;
-	public Integer ammount;
+	public long ammount;
 	
 	public DecisionInform() {
 		super("decisionInform");
